// apps/legacy-hono-api/src/v1/schema.ts
export const customerResponseSchema = z
  .object({
    customerId: z.string(),
    name: z.string().nullable(),
    email: z.string().nullable(),
    appUserId: z.string().nullable()
    // origin: z.enum(["dashboard", "ios", "android", "stripe", "api"]),
  })
  .meta({
    ref: 'Customer'
  });
// ...
export const sdkCustomerResponseSchema = z
  .object({
    customerId: z.string(),
    name: z.string().nullable(),
    email: z.string().nullable(),
    appUserId: z.string().nullable()
  })
  .meta({
    ref: 'SdkCustomer'
  });

export const sdkIdentifyCustomerBodySchema = z
  .object({
    appUserId: z.string(),
    name: z.string().optional(),
    email: z.string().email().optional()
  })
  .meta({
    ref: 'SdkIdentifyCustomerBody'
  });